shooting!

+++ src/main/java/components/physics/Rigidbody.java
@@ -1,7 +1,7 @@
 package components.physics;
  public abstract class Rigidbody {
      protected int mass=0;
-     protected int drag=0;
+     protected double drag=0;
      protected int xp;
      protected int yp;
      protected int v=0;
